tests/test_cycler.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-
#
# Copyright 2013-2019 European Commission (JRC);
# Licensed under the EUPL (the 'Licence');
# You may not use this work except in compliance with the Licence.
# You may obtain a copy of the Licence at: http://ec.europa.eu/idabc/eupl
import pandas as pd
import pytest

from wltp import cycler
from wltp.vehicle import calc_default_resistance_coeffs


def test_flatten_columns():
    cols = pd.MultiIndex.from_tuples([("a", "aa"), ("b", "")], names=("gear", "item"))
    fcols = cycler.flatten_cycle_columns(cols)
    infcols = cycler.inflate_cycle_columns(fcols)
    assert cols.equals(infcols)
    assert cols.names == infcols.names
    with pytest.raises(AssertionError, match="MultiIndex?"):
        cycler.inflate_cycle_columns(cols)
